moved variant matcher into own file
