+++ CMakeLists.txt
@@ -51,6 +51,7 @@ add_library(refl-cpp INTERFACE include/refl-cpp/refl-cpp.hpp
 
         include/refl-cpp/impl/impl.hpp
         include/refl-cpp/impl/variant.hpp
+        include/refl-cpp/impl/variant_matcher.hpp
 )
 target_link_libraries(refl-cpp INTERFACE
     fmt::fmt
